Abstracted out account edit form
The account edit form is now an extension of the datastore API, since
each account type is datastore specific.  The edit_form(account)
function in the datastore API will allow rendering of this datastore
specific form on person page without hardcoding the forms into the page.

This patch also adds a specialised template rendering tag that allows
the account object to return an unrendered template.  That is then
rendered using the {% inline_render form %} template syntax.

Change-Id: I64c7df9d51097fb81939913d8bc9722f10b98b3a